+++ packages/cosmos-ts/src/index.ts
@@ -1,0 +1,4 @@
+export * from './KeplrWallet'
+export * from './CosmJsWallet'
+export * from './types'
+export * from './endpoints'