+++ packages/token-metadata/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.22](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/token-metadata@0.0.21...@injectivelabs/token-metadata@0.0.22) (2021-10-25)
+
+
+### Features
+
+* added cosmos-ts package ([777893c](https://github.com/InjectiveLabs/injective-ts/commit/777893c6d278b37fc08a3635f2431cfa26902041))
+
+
+
+
+
 ## [0.0.21](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/token-metadata@0.0.20...@injectivelabs/token-metadata@0.0.21) (2021-10-20)